# [rust] Functional rewrite of top5 extraction