Could you give this a spin? I see a 4% improvement on my bare-metal machine, but not sure how well this performs in an Azure VM.