@jinyus 2023-11-20
A tad faster 3.16 vs 3.17 and dethrones go as the fastest for 5k posts.

Main: `22.07 ms | 316.11 ms | 2.84 s`

This PR:
```
Rust:

        Processing time (w/o IO): 21.66842ms
        total: 0.03s memory: 16660k
        Processing time (w/o IO): 21.552418ms
        total: 0.03s memory: 16524k
        Processing time (w/o IO): 22.852449ms
        total: 0.03s memory: 16404k
        Processing time (w/o IO): 21.597718ms
        total: 0.03s memory: 16404k
        Processing time (w/o IO): 21.695721ms
        total: 0.03s memory: 16532k
        Processing time (w/o IO): 21.622019ms
        total: 0.04s memory: 16532k
        Processing time (w/o IO): 21.62802ms
        total: 0.03s memory: 16528k
        Processing time (w/o IO): 21.597419ms
        total: 0.03s memory: 16404k
        Processing time (w/o IO): 21.705222ms
        total: 0.04s memory: 16396k
        Processing time (w/o IO): 21.582219ms
        total: 0.03s memory: 16532k

Rust:

        Processing time (w/o IO): 313.219025ms
        total: 0.37s memory: 39312k
        Processing time (w/o IO): 319.64448ms
        total: 0.38s memory: 39308k
        Processing time (w/o IO): 312.907618ms
        total: 0.37s memory: 39184k

Rust:

        Processing time (w/o IO): 2.82356714s
        total: 3.04s memory: 110660k
        Processing time (w/o IO): 2.817547507s
        total: 3.06s memory: 110416k
        Processing time (w/o IO): 2.839698157s
        total: 3.07s memory: 110420k

```